agent_layer/requirements.txt
langchain
openai
requests
python-dotenv
google-auth-oauthlib
langchain_community
